requirements.txt
confusable-homoglyphs==3.0.0
dj-database-url==0.4.2
Django==2.0.2
django-dotenv==1.4.2
django-extensions==1.9.9
django-rest-framework==0.1.0
djangorestframework==3.7.7
gunicorn==19.7.1
Pillow==5.0.0
psycopg2==2.7.3.2
pytz==2017.3
six==1.11.0
typing==3.6.4
whitenoise==3.3.1
